+++ lihil/interface/marks.py
@@ -1,27 +1,11 @@
 import re
-from types import GenericAlias
-from typing import (
-    Annotated,
-    Any,
-    AsyncGenerator,
-    Generator,
-    Literal,
-    TypeGuard,
-    TypeVar,
-    get_args,
-)
-from typing import get_origin as ty_get_origin
+from typing import Annotated, Any, AsyncGenerator, Generator, Literal, TypeVar
 
 from msgspec import Struct as Struct
-from typing_extensions import TypeAliasType
-
-from lihil.constant.status import Status
 
 T = TypeVar("T")
 
 LIHIL_RESPONSE_MARK = "__LIHIL_RESPONSE_MARK"
-LIHIL_PARAM_MARK = "__LIHIL_PARAM_MARK"
-LIHIL_PARAM_PATTERN = re.compile(r"__LIHIL_PARAM_MARK_(.*?)__")
 LIHIL_RETURN_PATTERN = re.compile(r"__LIHIL_RESPONSE_MARK_(.*?)__")
 
 
@@ -43,7 +27,6 @@ def extract_resp_type(mark: Any) -> "ResponseMark | None":
     return None
 
 
-
 # ================ Response ================
 
 TEXT_RETURN_MARK = resp_mark("text")
@@ -51,7 +34,6 @@ HTML_RETURN_MARK = resp_mark("html")
 STREAM_RETURN_MARK = resp_mark("stream")
 JSON_RETURN_MARK = resp_mark("json")
 EMPTY_RETURN_MARK = resp_mark("empty")
-JW_TOKEN_RETURN_MARK = resp_mark("jw_token")
 
 
 Text = Annotated[str | bytes, TEXT_RETURN_MARK, "text/plain"]
@@ -62,15 +44,5 @@ Stream = Annotated[
     "text/event-stream",
 ]
 Json = Annotated[T, JSON_RETURN_MARK, "application/json"]
-
-
-class Resp:
-    """
-    async def create_user() -> Annotated[Json[str], Resp(200)]
-    """
-
-    def __init__(self, code: Status):
-        self.code = code
-
 
 ResponseMark = Literal["text", "html", "stream", "json", "empty", "jw_token"]